+++ src/azure-cli-core/HISTORY.rst
@@ -6,6 +6,7 @@ Release History
 2.0.28
 ++++++
 * Fix issue that required extension to use `client_arg_name` keyword argument. This is no longer necessary.
+* Allow extensions to send telemetry with custom instrumentation key
 
 2.0.27
 ++++++